requirements.txt
streamlit
pandas
matplotlib
python-pptx
requests
sqlalchemy
psycopg2-binary
python-dotenv